Fixed 2 bugs.
Fixed a bug where the app would crash if the menu action was performed while in the History View. Also fixed a bug on non-touch-screen devices where if you select the History Button to go to the History View and go back, you cannot immeadiately select the History Button again.